Feat: Add OutMessagePacket support for PacketEvents [desc]
Also began implementing ProtocolPlatform instance for PacketEvents.

Made some adjustments to PreChatPacketSendEvent, as that directly references ProtocolLib PacketContainers, which isn't ideal as PE code also needs to trigger that event.

We've thought of making a wrapper but more thought needs to be invested concerning what information would actually be stored within that wrapper that would actually be useful (after all, it's in the API so it needs to be relatively usable), as well as needing to figure out a nice way to map the PacketWrapper<?>(s) (PE) and PacketContainer(s) (PLib) to this wrapper.