Chore: PacketEvents adventure API pain [desc]
Thanks to the FANTASTIC decision of PacketEvents to shade Adventure API and NOT shade it, we encounter a large amount of errors, where we try to provide our own instance of Adventure (which is shaded and relocated), which obviously won't work.

We can't have any helper classes that convert to PE's shaded adventure because it's shaded but not relocated, so any instance of it will just be relocated by the compiler to our shaded version.

In the end, we do a bunch of reflection fuckery.